remove circular dep between hidden power  and abilities
