+++ app/core/abilities/explosion.ts
@@ -1,0 +1,42 @@
+import { AttackType } from "../../types/enum/Game"
+import { Item } from "../../types/enum/Item"
+import type { Board } from "../board"
+import type { PokemonEntity } from "../pokemon-entity"
+import { AbilityStrategy } from "./ability-strategy"
+
+class ExplosionStrategy extends AbilityStrategy {
+  process(
+    pokemon: PokemonEntity,
+    board: Board,
+    target: PokemonEntity,
+    crit: boolean
+  ) {
+    super.process(pokemon, board, target, crit)
+    const damage = [50, 100, 200][pokemon.stars - 1] ?? 200
+    const cells = board.getAdjacentCells(pokemon.positionX, pokemon.positionY)
+
+    cells.forEach((cell) => {
+      if (cell.value && pokemon.team != cell.value.team) {
+        cell.value.handleSpecialDamage(
+          damage,
+          board,
+          AttackType.SPECIAL,
+          pokemon,
+          crit
+        )
+      }
+    })
+
+    if (!pokemon.items.has(Item.PROTECTIVE_PADS)) {
+      pokemon.handleSpecialDamage(
+        damage,
+        board,
+        AttackType.SPECIAL,
+        pokemon,
+        crit
+      )
+    }
+  }
+}
+
+export const explosionStrategy = new ExplosionStrategy()
