+++ app/core/abilities/hidden-power.ts
@@ -4,7 +4,6 @@ import PokemonFactory from "../../models/pokemon-factory"
 import { getPokemonData } from "../../models/precomputed/precomputed-pokemon-data"
 import { PRECOMPUTED_POKEMONS_PER_TYPE_AND_CATEGORY } from "../../models/precomputed/precomputed-types-and-categories"
 import type { IPokemon } from "../../types"
-import { Ability } from "../../types/enum/Ability"
 import { AttackType, Rarity } from "../../types/enum/Game"
 import {
   Berries,
@@ -21,8 +20,11 @@ import { pickNRandomIn, pickRandomIn, randomWeighted } from "../../utils/random"
 import type { Board } from "../board"
 import { giveRandomEgg } from "../eggs"
 import type { PokemonEntity } from "../pokemon-entity"
-import { castAbility } from "./abilities"
 import { AbilityStrategy } from "./ability-strategy"
+import { castAbility } from "./cast"
+import { explosionStrategy } from "./explosion"
+import { meditateStrategy } from "./meditate"
+import { thunderShockStrategy } from "./thunder-shock"
 
 export class HiddenPowerStrategy extends AbilityStrategy {
   requiresTarget = false
@@ -233,7 +235,7 @@ export class HiddenPowerNStrategy extends HiddenPowerStrategy {
           const target = board.getEntityOnCell(pokemon.targetX, pokemon.targetY)
           if (target) {
             pokemon.addShield(50, unown, 1, false)
-            castAbility(Ability.EXPLOSION, pokemon, board, target, false)
+            castAbility(explosionStrategy, pokemon, board, target, false)
           }
         }
       }
@@ -372,7 +374,7 @@ export class HiddenPowerVStrategy extends HiddenPowerStrategy {
     super.process(unown, board, target, crit)
     board.forEach((x: number, y: number, enemy: PokemonEntity | undefined) => {
       if (enemy && unown.team !== enemy.team) {
-        castAbility(Ability.THUNDER_SHOCK, unown, board, enemy, false)
+        castAbility(thunderShockStrategy, unown, board, enemy, false)
       }
     })
   }
@@ -450,7 +452,7 @@ export class HiddenPowerYStrategy extends HiddenPowerStrategy {
     super.process(unown, board, target, crit)
     board.forEach((x: number, y: number, ally: PokemonEntity | undefined) => {
       if (ally && unown.team === ally.team) {
-        castAbility(Ability.MEDITATE, ally, board, ally, false)
+        castAbility(meditateStrategy, ally, board, ally, false)
       }
     })
   }
