+++ app/core/attacking-state.ts
@@ -6,7 +6,8 @@ import { EffectEnum } from "../types/enum/Effect"
 import { PokemonActionState } from "../types/enum/Game"
 import { distanceC } from "../utils/distance"
 import { max } from "../utils/number"
-import { castAbility } from "./abilities/abilities"
+import { AbilityStrategies } from "./abilities/abilities"
+import { castAbility } from "./abilities/cast"
 import type { Board } from "./board"
 import type { PokemonEntity } from "./pokemon-entity"
 import PokemonState from "./pokemon-state"
@@ -87,7 +88,7 @@ export default class AttackingState extends PokemonState {
 
         if (pokemon.pp >= pokemon.maxPP && pokemon.canCast) {
           // CAST ABILITY
-          castAbility(pokemon.skill, pokemon, board, target)
+          castAbility(AbilityStrategies[pokemon.skill], pokemon, board, target)
         } else {
           // BASIC ATTACK
           pokemon.count.attackCount++
