+++ app/core/moving-state.ts
@@ -5,7 +5,8 @@ import { Passive } from "../types/enum/Passive"
 import { Synergy } from "../types/enum/Synergy"
 import { distanceC } from "../utils/distance"
 import { findPath } from "../utils/pathfind"
-import { AbilityStrategies, castAbility } from "./abilities/abilities"
+import { AbilityStrategies } from "./abilities/abilities"
+import { castAbility } from "./abilities/cast"
 import type { Board } from "./board"
 import { OnMoveEffect } from "./effects/effect"
 import { drumBeat, partingShot, stenchJump } from "./effects/passives"
@@ -41,7 +42,7 @@ export default class MovingState extends PokemonState {
         pokemon.canCast &&
         AbilityStrategies[pokemon.skill]?.requiresTarget === false
       ) {
-        castAbility(pokemon.skill, pokemon, board, null)
+        castAbility(AbilityStrategies[pokemon.skill], pokemon, board, null)
       } else if (targetAtRange) {
         pokemon.toAttackingState()
       } else if (
